src/commands/utility/cleanup/requirements.txt
bcrypt==5.0.0
cffi==2.0.0
cryptography==46.0.3
invoke==2.2.1
paramiko==4.0.0
psycopg==3.3.1
psycopg-binary==3.3.1
pycparser==2.23
PyNaCl==1.6.1
python-dotenv==1.2.1
scp==0.15.0
typing_extensions==4.15.0
